Restructure /api
