+++ src/api/movieApi/endpoints/fetchMovieDetails.ts
@@ -1,10 +1,10 @@
-import type {Movie} from './types';
-import {movieApiService} from './movieApiService';
+import type {Movie} from '../types';
+import {client} from '../client';
 import type {AxiosError} from 'axios';
 
 export const fetchMovieDetails = async (movieId: string): Promise<Movie> => {
   try {
-    return (await movieApiService.get(`/movie/${movieId}`)).data;
+    return (await client.get(`/movie/${movieId}`)).data;
   } catch (error) {
     const axiosError = error as AxiosError<{status_message: string}>;
     throw {
