+++ src/pages/MovieDetailsPage/MovieDetailsPage.tsx
@@ -1,8 +1,8 @@
 import {useQuery, useQueryClient} from '@tanstack/react-query';
 import {useParams} from 'react-router-dom';
 import {format} from 'date-fns';
-import {fetchMovieDetails} from '../../api/movieApi/fetchMovieDetails';
-import type {ApiErrorResponseDetails, Movie} from '../../api/movieApi/types';
+import {fetchMovieDetails} from '../../api/movieApi/endpoints/fetchMovieDetails';
+import type {ApiErrorResponseDetails} from '../../api/movieApi/utils/types';
 import ApiErrorScreen from '../../app/components/ApiErrorScreen/ApiErrorScreen';
 import PopcornLoader from '../common/PopcornLoader/PopcornLoader';
 import {
@@ -21,8 +21,9 @@ import {
 } from './MovieDetailsPage.styled';
 import MovieCover from '../common/MovieCover/MovieCover';
 import GoBackLink from './components/GoBackLink/GoBackLink';
-import {getErrorFallbackMessage} from '../../api/movieApi/utils/getErrorFallbackMessage.ts';
+import {getErrorFallbackMessage} from '../../api/movieApi/utils/getErrorFallbackMessage';
 import {API_STATUS_MESSAGE, ERRORS_BY_HTTP_STATUS} from '../../api/movieApi/constants';
+import type {Movie} from '../../api/movieApi/types';
 
 export default function MovieDetailsPage() {
   const queryClient = useQueryClient();
@@ -53,7 +54,7 @@ export default function MovieDetailsPage() {
       <ApiErrorScreen
         errorMessage={
           error
-            ? getErrorFallbackMessage(error.status, error.message)
+            ? getErrorFallbackMessage({status: error.status, message: error.message})
             : ERRORS_BY_HTTP_STATUS[404][API_STATUS_MESSAGE.RESOURCE_NOT_FOUND]
         }
         onReset={() => {
